fix: fix next build, update project structure, use google font
