+++ .prettierignore
@@ -1,6 +1,4 @@
 node_modules
 .next
 dist
-
-*.min.css
-*.min.js
+public
