+++ .prettierrc
@@ -20,6 +20,7 @@
     "^@/hooks(/.*)?$",
 
     "",
+    "^@/providers(/.*)?$",
     "^@/components/ui(/.*)?$",
     "^@/components(/.*)?$",
     "^@/sections(/.*)?$",
