+++ src/app/(action)/layout.tsx
@@ -3,12 +3,15 @@ import React from 'react'
 import { leftSidebarStudentData } from '@/configs'
 
 import DashboardLayout from '@/components/layouts/DashboardLayout'
+import ProtectedRoute from '@/components/shared/protected-route'
 
 const layout = ({ children }: { children: React.ReactNode }) => {
   return (
-    <DashboardLayout leftSidebarData={leftSidebarStudentData}>
-      {children}
-    </DashboardLayout>
+    <ProtectedRoute>
+      <DashboardLayout leftSidebarData={leftSidebarStudentData}>
+        {children}
+      </DashboardLayout>
+    </ProtectedRoute>
   )
 }
