+++ src/app/course/layout.tsx
@@ -1,16 +1,7 @@
-import React from 'react'
-
-import { manropeFont } from '@/components/common/fonts'
-import QueryProvider from '@/components/layouts/QueryProvider'
+import QueryProvider from '@/providers/query-provider'
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <html lang="en">
-      <body className={`${manropeFont.className} antialiased`}>
-        <QueryProvider>{children}</QueryProvider>
-      </body>
-    </html>
-  )
+  return <QueryProvider>{children}</QueryProvider>
 }
 
 export default Layout
