+++ src/assets/css/themes.css
@@ -1,0 +1,9 @@
+@import '../../../public/assets/css/bootstrap.css';
+@import '../../../public/assets/css/animate.min.css';
+@import '../../../public/assets/css/mmenu.css';
+@import '../../../public/assets/css/swiper-bundle.min.css';
+@import '../../../public/assets/css/magnific-popup.min.css';
+@import '../../../public/assets/css/styles.css';
+@import '../../../public/assets/font/fonts.css';
+@import '../../../public/assets/icons/flat/flaticon_upskill.css';
+@import '../../../public/assets/icons/icomoon/style.css';
