+++ src/components/layouts/CommonLayout.tsx
@@ -1,16 +1,5 @@
-import { ToastContainer } from 'react-toastify'
-
-import { manropeFont } from '@/components/common/fonts'
-import Header from '@/components/themes/Header'
-import HeadLinks from '@/components/themes/HeadLinks'
-import HeadScripts from '@/components/themes/HeadScripts'
-
+import Header from '../themes/Header'
 import Footer from './Footer'
-import QueryProvider from './QueryProvider'
-
-import '@/components/themes/assetsImports'
-
-// import Script from 'next/script'
 
 interface LayoutProps {
   children?: React.ReactNode
@@ -18,31 +7,11 @@ interface LayoutProps {
 
 const CommonLayout = ({ children }: LayoutProps) => {
   return (
-    <html lang="en">
-      <HeadLinks />
-      <body className={`${manropeFont.className} counter-scroll antialiased`}>
-        <div id="wrapper">
-          <QueryProvider>
-            <Header />
-            <main>{children}</main>
-            <Footer />
-          </QueryProvider>
-        </div>
-      </body>
-      <HeadScripts />
-      <ToastContainer
-        position="top-right"
-        autoClose={5000}
-        hideProgressBar={false}
-        newestOnTop
-        closeOnClick={false}
-        rtl={false}
-        pauseOnFocusLoss
-        draggable
-        pauseOnHover
-        theme="light"
-      />
-    </html>
+    <>
+      <Header />
+      <main>{children}</main>
+      <Footer />
+    </>
   )
 }
