+++ src/components/shared/protected-route.tsx
@@ -1,0 +1,24 @@
+'use client'
+
+import { useEffect } from 'react'
+import { useRouter } from 'next/navigation'
+import { useAuthStore } from '@/stores/useAuthStore'
+
+interface Props {
+  children: React.ReactNode
+}
+
+const ProtectedRoute: React.FC<Props> = ({ children }) => {
+  const { isAuthenticated } = useAuthStore()
+  const router = useRouter()
+
+  useEffect(() => {
+    if (!isAuthenticated) {
+      router.replace('/sign-in')
+    }
+  }, [isAuthenticated, router])
+
+  return isAuthenticated ? <>{children}</> : null
+}
+
+export default ProtectedRoute
