+++ src/components/themes/HeadLinks.tsx
@@ -1,23 +1,25 @@
-import React from 'react'
-import Head from 'next/head'
-import Script from 'next/script'
-
 const HeadLinks = () => {
   return (
     <>
-      <Head>
+      {/* <Head>
         <link rel="shortcut icon" href="/images/logo/favicon.png" />
         <link
           rel="apple-touch-icon-precomposed"
           href="/images/logo/favicon.png"
         />
-      </Head>
-      <Script
+      </Head> */}
+      {/* <Script
         src="https://code.jquery.com/jquery-3.6.0.min.js"
         strategy="afterInteractive"
         integrity="sha384-KyZXEAg3QhqLMpG8r+Knujsl5+5hb7V6Wo3E2s8Zr1An1pF7XKp5yzl5tD6U4sOY"
         crossOrigin="anonymous"
-      />
+      /> */}
+      {/* <Script
+        src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.6.0/jquery.min.js"
+        integrity="sha512-894YE6QWD5I59HgZOGReFYm4dnWc1Qt5NtvYSaNcOP+u1T9qYdvdihz0PPSiiqn/+/3e7Jo4EaG7TubfWGUrMQ=="
+        crossOrigin="anonymous"
+        referrerPolicy="no-referrer"
+      /> */}
     </>
   )
 }
