--- src/components/themes/assetsImports.ts
@@ -1,11 +1,0 @@
-import '../../../public/assets/css/bootstrap.css'
-import '../../../public/assets/css/animate.min.css'
-import '../../../public/assets/css/mmenu.css'
-import '../../../public/assets/css/swiper-bundle.min.css'
-import '../../../public/assets/css/magnific-popup.min.css'
-import '../../../public/assets/css/styles.css'
-import '../../../public/assets/font/fonts.css'
-import '../../../public/assets/icons/flat/flaticon_upskill.css'
-import '../../../public/assets/icons/icomoon/style.css'
-
-// export default {}
