+++ src/sections/home/components/course-list.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import { useEffect, useState } from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
