+++ src/sections/home/components/post-list.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import Image from 'next/image'
 import Link from 'next/link'
 import { Pagination } from 'swiper/modules'
