+++ src/sections/lessons/view/learning-lesson-view.tsx
@@ -1,4 +1,4 @@
-'use client'
+// 'use client'
 
 import React from 'react'
 import Image from 'next/image'
